+++ Source/Core/Core/IOS/USB/Bluetooth/WiimoteDevice.h
@@ -9,7 +9,6 @@
 #include <string>
 
 #include "Common/CommonTypes.h"
-#include "Common/Swap.h"
 #include "Core/IOS/USB/Bluetooth/hci.h"
 
 class PointerWrap;
@@ -22,22 +21,6 @@ namespace Device
 {
 class BluetoothEmu;
 }
-
-class CBigEndianBuffer
-{
-public:
-  CBigEndianBuffer(u8* pBuffer) : m_pBuffer(pBuffer) {}
-  u8 Read8(u32 offset) const { return m_pBuffer[offset]; }
-  u16 Read16(u32 offset) const { return Common::swap16(*(u16*)&m_pBuffer[offset]); }
-  u32 Read32(u32 offset) const { return Common::swap32(*(u32*)&m_pBuffer[offset]); }
-  void Write8(u32 offset, u8 data) { m_pBuffer[offset] = data; }
-  void Write16(u32 offset, u16 data) { *(u16*)&m_pBuffer[offset] = Common::swap16(data); }
-  void Write32(u32 offset, u32 data) { *(u32*)&m_pBuffer[offset] = Common::swap32(data); }
-  u8* GetPointer(u32 offset) { return &m_pBuffer[offset]; }
-
-private:
-  u8* m_pBuffer;
-};
 
 class WiimoteDevice
 {
